sdk: update code to sdk release v1.3.1
WLAN:
* Fixed compatibility problems with the network assistant unable to lock the channel
* Added interface for generating 32-byte psk based on ssid, passphrase

CedarX & Audio decoder
* Support AAC SBR, m3u8, HTTPS (libcedarx_pro.a)
* support play online audio stream
* Memory save
* Fix playback m4a memory leak problem
* Solve several bugs

Driver
* Fix adc channel 8 detection voltage is not accurate
* Fix SPI crossover error
* Disable watchdog interrupt mode by default
* Several driver DMA configuration optimization
* XR809 flash support

General
* Memory save
* Support time(), gettimeofday(), support time zone settings
* Support for generating mac addresses based on chip id
* Sound wave optimization (the new method is not compatible with the old version)
* Solve the problem of OTA file upgrade and pull out SD card crash
* Simplify dhcp server configuration implementation

Tool
* mkimage adds extensions: generate img files with OTA information
* phoenixMC: update the linux version to resolve some compatibility issues
* oem_cli_tool: Fixed a problem when the flash read length was between 128-1024
* Voiceprint distribution network apk: new data packet transmission network distribution method
* smartconfig apk: the port number of receiving UDP packets is changed from 10000 to 10001

Documentation
* XR871声波配网升级指导.pdf
* XR871_RF_Validation_Report_V1.0-EN

===============================================================
* WiFi
* 修复配网助手无法锁定信道的兼容性问题
* 增加接口用于根据ssid、passphrase生成32-byte psk

* 播控
* 支持AAC SBR、m3u8、HTTPS(libcedarx_pro.a)
* 增加queue stream以支持音频数据流
* 内存优化
* 修复播放m4a内存泄漏问题
* 解决若干bug

* 驱动
* 修复adc channel 8检测电压不准的问题
* 修复SPI分频不准问题
* disable watchdog interrupt mode by default
* 若干驱动DMA配置优化
* XR809 SIP flash型号的支持

* 综合
* 内存优化
* 支持time(), gettimeofday()，支持时区设置
* 支持根据chip id生成mac地址
* 声波配网优化(优化方案与旧方案不兼容)
* 解决OTA file升级拔出SD卡死机的问题
* 精简dhcp server配置的实现方式

* 工具
* mkimage增加扩展功能：生成带有OTA信息的img文件
* phoenixMC：更新linux版本，解决部分兼容性问题
* oem_cli_tool：修复flash读取长度为128-1024之间时失败的问题
* 声纹配网apk：新增数据分组发送的配网方式
* smartconfig apk：接收UDP包的端口号由10000改为10001

* 文档
* XR871声波配网升级指导
* XR871_RF_Validation_Report_V1.0-EN

+++ 03_SDK/xr871sdk/src/driver/chip/sdmmc/sd.c
@@ -707,7 +707,7 @@ free_card:
 }
 
 #ifdef CONFIG_SD_PM
-static uint32_t mmc_sd_suspending;
+static uint8_t mmc_sd_suspending;
 
 static int32_t mmc_sd_suspend(struct mmc_host *host)
 {
